+++ src/scribe_data/cli/cli_utils.py
@@ -22,7 +22,7 @@ Utility functions for the Scribe-Data CLI.
 
 import json
 from pathlib import Path
-from typing import Dict, List, Union
+from typing import Union
 
 LANGUAGE_METADATA_FILE = (
     Path(__file__).parent.parent / "resources" / "language_metadata.json"
@@ -40,6 +40,11 @@ with DATA_TYPE_METADATA_FILE.open("r", encoding="utf-8") as file:
 
 language_map = {
     lang["language"].lower(): lang for lang in language_metadata["languages"]
+}
+
+# Create language_to_qid dictionary
+language_to_qid = {
+    lang["language"].lower(): lang["qid"] for lang in language_metadata["languages"]
 }
 
 
@@ -66,7 +71,7 @@ def correct_data_type(data_type: str) -> str:
             return wt
 
 
-def print_formatted_data(data: Union[Dict, List], data_type: str) -> None:
+def print_formatted_data(data: Union[dict, list], data_type: str) -> None:
     """
     Prints a formatted output from the Scribe-Data CLI.
     """